moved test files in __tests__
